Add auth services
This adds a /session path that can accept POST, GET, and DELETE
requests.
* POST will login, assuming that the credentials are correct
* GET will return the currently logged-in user's user representation
  when a user is logged in and a 401 otherwise
* DELETE will logout

The POST requests will accept a json with keys username: and password:,
the GET request will return the username: only and delete needs no
additional information.

All of the auth is handled by friend and hangs on Ring's session cookie,
this should be considered an implicit parameter in /session requests and
any other requests that require the user to be authed.