Add the beginnings of the front end webpage
This change brings in some javascript, backed by react.js for views and
jQuery for ajax calls. Views have been added for the login page and for
the expenses page.

The login page is implemented with a form that calls back to the server
to perform auth, the expenses page is not implemented yet and, so far,
only contains a logout link.

Of the javascripts:
* app.js is our main app
* login.js handles all calls and callbacks related to auth
* expense.js handles all calls and callbacks related to expenses

* require.js is a 3rd party lib

Of the css:
* app.css is ours and will contain the styling for the full app
* normalize.3.0.2.css is 3rd party and standardizes defaults over all
  browsers.