Replace in memory datastore with Postgres
This is a wide and varied change which adds:
* A db migrations folder + script
* An SQL script for creating a user table
* Environment variables for db connection

And it also updates the `users.clj` to use the Postgres via a yesql
interface (see users.sql), fully replacing the atom of app state.

The database config is added in the top (i.e. outer) layer,
handler.clj. That's not ideal since handler is not really the top of our
app but the top of the http stack, but since we're using lein ring then
that's just the way we're going to have to go.

This change now parameterises the route creation so that they can
locally bind on the database configuration

The environment variables now required for running the app are:
* EXPENSAPP_DBHOST
* EXPENSAPP_DBPORT
* EXPENSAPP_DBNAME
* EXPENSAPP_DBUSER
* EXPENSAPP_DBPASS